requirements.txt
flask
gunicorn
pandas
PyYAML
scikit-learn
numpy
